+++ core/src/main/java/com/ibm/wala/analysis/stackMachine/AbstractIntStackMachine.java
@@ -528,7 +528,7 @@ public abstract class AbstractIntStackMachine implements FixedPointConstants {
   }
 
   /** Representation of the state of the JVM stack machine at some program point. */
-  public class MachineState extends AbstractVariable<MachineState> {
+  public final class MachineState extends AbstractVariable<MachineState> {
     private int[] stack;
 
     private int[] locals;