+++ dalvik/src/main/java/com/ibm/wala/dalvik/analysis/typeInference/DalvikTypeVariable.java
@@ -3,7 +3,7 @@ package com.ibm.wala.dalvik.analysis.typeInference;
 import com.ibm.wala.analysis.typeInference.TypeAbstraction;
 import com.ibm.wala.analysis.typeInference.TypeVariable;
 
-public class DalvikTypeVariable extends TypeVariable {
+public final class DalvikTypeVariable extends TypeVariable {
   private final boolean isIntZeroConstant;
 
   public DalvikTypeVariable(TypeAbstraction type, boolean isIntZeroConstant) {