+++ util/src/main/java/com/ibm/wala/util/intset/EmptyIntSet.java
@@ -13,7 +13,7 @@ package com.ibm.wala.util.intset;
 import java.io.Serial;
 import java.util.NoSuchElementException;
 
-public class EmptyIntSet implements IntSet {
+public final class EmptyIntSet implements IntSet {
 
   @Serial private static final long serialVersionUID = 5116475799916663164L;
   public static EmptyIntSet instance = new EmptyIntSet();